feat: Add author and publication date to InfoPublication
The code changes in this commit add the author and publication date to the InfoPublication component. The `InfoPublication` component now displays the author's name and the publication date below the content. This enhancement improves the user experience by providing additional information about the publication.

Recent user commits:
- Publications have been stylized
- Publications can be showed
- Home Page created
- Login stylized
- Login is working

Recent repository commits:
- Publications have been stylized
- Publications can be showed
- Home Page created
- Login stylized
- Login is working
- Initial commit